rewrote MPI stuff to use dtensor data structure
A dtensor is an ordered tuple of triplets (n, ib, ob) giving the size of
a dimension (n) and its input and output block sizes of a distributed
row-major multi-dimensional array.  An MPI DFT (etc.) is now specified
in terms of dtensors, which provide a much more flexible data layout.

For example, we can now describe multidimensional block distributions,
which are important if the number of processors is greater than the
size of any given dimension.  Currently, we only have solvers for
1d slab distributions, and this is all that is supported in the basic
and advanced APIs.  The guru API allows one to specify more general
distributions, however, which will be useful when/if we have solvers
for this case.

We now also don't need a TRANSPOSED flag, at least internally, since
TRANSPOSED multi-dimensional DFT plans just correspond to dtensors
where the input and output block distributions are different.

Other changes include the use of the XM(foo) macro for X(mpi_foo).