Fix bug in bubblesort
Bubblesort was not sorting.  This was a bug in the benchmark library
(not in FFTW per se), and it impacted the benchmark program
with --report-time and --report-mflops causing it to output
an incorrect value for the median.  (The minimum, maximum, and
average value were correct.)  Thanks Dima Baksheev of Intel for
reporting this bug.